backend/requirements.txt
fastapi
uvicorn
scikit-learn
xgboost
pandas
requests
beautifulsoup4
python-multipart
tldextract
python-levenshtein
numpy
